Move `DNSHandlerAdapter` implementations to separate package; fix `RequiredIfUnset()` for aliases
