+++ pkg/controller/provider/aws/config/config.go
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: SAP SE or an SAP affiliate company and Gardener contributors
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package config
+
+type AWSConfig struct {
+	BatchSize int `json:"batchSize"`
+}
