+++ pkg/controller/provider/azure-private/factory.go
@@ -10,7 +10,7 @@ import (
 	"github.com/gardener/external-dns-management/pkg/dns/provider"
 )
 
-const TYPE_CODE = "azure-private-dns"
+const TYPE_CODE = utils.ProviderTypeAzurePrivateDNS
 
 var rateLimiterDefaults = provider.RateLimiterOptions{
 	Enabled: true,
@@ -18,7 +18,7 @@ var rateLimiterDefaults = provider.RateLimiterOptions{
 	Burst:   10,
 }
 
-var Factory = provider.NewDNSHandlerFactory(TYPE_CODE, NewHandler, utils.NewAdapter(TYPE_CODE)).
+var Factory = provider.NewDNSHandlerFactory(NewHandler, utils.NewAdapter(TYPE_CODE)).
 	SetGenericFactoryOptionDefaults(provider.GenericFactoryOptionDefaults.SetRateLimiterOptions(rateLimiterDefaults))
 
 func init() {
